requirements.txt
requests
python-dateutil
prometheus_client
flask
